2003-09-09 03:20 UTC+0200 Przemyslaw Czerpak <druzus@polbox.com>   * xharbour/include/simplex.c   * xharbour/source/compiler/genc.c   * xharbour/source/compiler/hbusage.c     * casting and some nonstandard C statements corrected
  * xharbour/source/pp/pptable.c
    + added some missing CLIPPER 5.3 statements:
      SET MBLOCKSIZE TO <x>
      SET MEMOBLOCK TO <x>
      SET MFILEEXT TO <x>
      SET AUTOSHARE TO [<x>]
      SET AUTORDER TO [<x>]
      SET STRICTREAD <x:ON,OFF,&>

  * xharbour/source/rdd/dbffpt/dbffpt1.c
    + DBFFPT now uses _SET_MFILEEXT and _SET_MBLOCKSIZE

+++ xharbour/source/compiler/genc.c
@@ -1,5 +1,5 @@
 /*
- * $Id: genc.c,v 1.48 2003/07/15 09:15:33 andijahja Exp $
+ * $Id: genc.c,v 1.49 2003/07/21 22:41:06 ronpinkas Exp $
  */
 
 /*
@@ -1590,7 +1590,7 @@ static HB_GENC_FUNC( hb_p_pushdouble )
 
    fprintf( cargo->yyc, "\tHB_P_PUSHDOUBLE," );
    ++lPCodePos;
-   for( i = 0; i < sizeof( double ) + sizeof( BYTE ) + sizeof( BYTE ); ++i )
+   for( i = 0; i < ( int ) ( sizeof( double ) + sizeof( BYTE ) + sizeof( BYTE ) ); ++i )
       fprintf( cargo->yyc, " %i,", ( ( BYTE * ) pFunc->pCode )[ lPCodePos + i ] );
    if( cargo->bVerbose ) fprintf( cargo->yyc, "\t/* %.*f, %d, %d */",
       *( ( BYTE * ) &( pFunc->pCode[ lPCodePos + sizeof( double ) ] ) ),
