+++ xharbour/source/compiler/hbusage.c
@@ -1,5 +1,5 @@
 /*
- * $Id: hbusage.c,v 1.10 2003/06/23 14:04:15 andijahja Exp $
+ * $Id: hbusage.c,v 1.12 2003/07/24 20:53:45 andijahja Exp $
  */
 
 /*
@@ -105,7 +105,7 @@ void hb_compPrintUsage( char * szSelf )
    printf( "\nSyntax:  %s <file[s][.prg]> [options]"
            "\n", szSelf );
 
-   for( iLine = 0; iLine < ( sizeof( szOptions ) / sizeof( char * ) ); iLine++ )
+   for( iLine = 0; iLine < ( int ) ( sizeof( szOptions ) / sizeof( char * ) ); iLine++ )
       printf( szOptions[ iLine ], OS_OPT_DELIMITER_LIST[ 0 ] );
 }
 
@@ -129,7 +129,7 @@ void hb_compPrintModes( void )
 
    printf( "\nCompatibility flags (lowercase/uppercase significant): -k[options]\n" );
 
-   for( iLine = 0; iLine < ( sizeof( szOptions ) / sizeof( char * ) ); iLine++ )
+   for( iLine = 0; iLine < ( int ) ( sizeof( szOptions ) / sizeof( char * ) ); iLine++ )
       printf( szOptions[ iLine ] );
 }
